+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning 2.0.0](https://semver.org/spec/
 ### Fixed
 
 - Comment count while getting posts
+- Implement JWT refresh tokens
 
 ## [2.0.0] - 2026-06-29